Place and remove blocks

+++ 3D Game Engine/3D Game Engine/Objects/MeshLoader.h
@@ -14,7 +14,7 @@
 typedef struct {
     Vertex*   vertices;
     uint32_t  vertexCount;
-    uint16_t* indices;
+    uint32_t* indices;
     uint32_t  indexCount;
 } MeshData;
